:hammer: Moved Student Home Page related files to new sub directory
